requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
langchain_community==0.3.3
langchain_core==0.3.12
langchain_google_genai==2.0.1
langgraph==0.2.39
Pillow==10.0.0
pydantic==2.9.2
python-dotenv==1.0.1
ratelimit==2.2.1
python-multipart==0.0.12
